+++ migrations/20200427230859-emails.js
@@ -1,7 +1,7 @@
 module.exports = {
   up: (queryInterface, Sequelize) => queryInterface.createTable('emails', {
     serviceID: {
-      type: Sequelize.INTEGER(11),
+      type: Sequelize.INTEGER,
       primaryKey: true,
       references: {
         model: 'services',
